refactor: wip - working path
#18
